AIFJAN3/ProjetAIF/requirements.txt
gradio==3.39.0
annoy==1.16.3
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
tqdm
scikit-learn
transformers
nltk
tensorflow